+++ action/textarea/textarea.js
@@ -3,15 +3,15 @@ import guid from 'guid';
 
 class Input extends App{
 
-    constructor() {
-        super('input');
+    constructor(tagName) {
+        super('textarea');
         this.guid = guid();
         // 初始化
-        this.attr({type: 'text', class: 'form-item form-control', guid: this.guid});
+        this.attr({class: 'form-area', guid: this.guid});
     }
 
     reset() {
-        this.$.val('');
+        this.$dom.val('');
         return this;
     }
 }