+++ changelog.md
@@ -1,3 +1,17 @@
+v3.0.0
+---
+* 删除``group``相关功能
+* 删除``form.apps``属性
+* 删除``form.getGroupname``
+* 删除``form.create``
+* 删除``form.group``
+* 删除``form.getApp``
+* 删除``form.css``
+* 删除``form.dataChange`` 改功能由``form.on('data', handler)``提供支持
+* 新增``form.on`` 支持消息订阅
+* 删除``app.create``
+* ``app.render`` 更名为 ``app.create``
+
 v2.0.0
 ---
 重写``form``，彻底改变其工作原理、渲染方式，使其更加易用